pyproject.toml
[tool.poetry]
name = "educational-website"
version = "0.1.0"
description = ""
authors = ["KamilRahmatullin <kamil.raxmatullin.01@bk.ru>"]
license = "MIT"
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
fastapi = "^0.111.0"
pytest = "^8.2.2"
uvicorn = {extras = ["standart"], version = "^0.30.1"}
sqlalchemy = {extras = ["asyncio"], version = "^2.0.30"}
aiosqlite = "^0.20.0"
pydantic-settings = "^2.3.1"
alembic = "^1.13.1"
pyjwt = {extras = ["crypto"], version = "^2.8.0"}
bcrypt = "^4.1.3"
python-multipart = "^0.0.9"


[tool.poetry.group.dev.dependencies]
black = "^24.4.2"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
